net/gve: support out-of-order completions on DQ Rx
The DPDK DQ driver made an implicit assumption that buffers will be
returned from the NIC in the order which they are posted. While this is
generally the case, there are certain edge cases, such as HW GRO, where
buffers may be completed out of order. In such cases, the driver would
have returned the wrong mbuf to the program, which would cause issues if
the mbuf is freed while the NIC still expects to own the buffer or the
application receives a buffer which contains garbage data.

This patch updates the Rx path behavior more in line with what is
expected by the NIC by utilizing the buf_id in the completion descriptor
to decide which buffer to complete. A stack is introduced to hold the
available buf_ids when posting to handle the fact that buf_ids can come
back from the hardware out of order.

Signed-off-by: Joshua Washington <joshwash@google.com>
Reviewed-by: Ankit Garg <nktgrg@google.com>