dpdk: support quick jump to API definition
Currently, the RTE_EXPORT_INTERNAL_SYMBOL, RTE_EXPORT_SYMBOL and
RTE_EXPORT_EXPERIMENTAL_SYMBOL are placed at the beginning of APIs,
but don't end with a semicolon. As a result, some IDEs cannot identify
the APIs and cannot quickly jump to the definition.

A semicolon is added to the end of above RTE_EXPORT_XXX_SYMBOL in this
commit.

And also change the gen-version-map.py to ensure it only identifies
RTE_EXPORT_XXX_SYMBOL that end with a semicolon.

Signed-off-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>