dts: allow mbuf_fast_free to be set with testpmd shell
Currently, there is no way in the testpmd shell class to set the mbuf
fast free offload configuration for RX or TX ports. This prohibits any
test suites to be written utilizing this offload configuration. Additionally,
the NIC capabilities for the TX offload are not gathered. This prevents future
test suites from being skipped if they do not support a TX offload capability.
Introduce methods that support calls to testpmd in order to allow the
configuration of mbuf fast free and to gather TX offload capabilities.

Signed-off-by: Andrew Bailey <abailey@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>