lib: add semicolon when export experimental symbol
Currently, the RTE_EXPORT_EXPERIMENTAL_SYMBOL was placed at the
beginning of APIs, but doesn't end with a semicolon. As a result, some
IDEs cannot identify the APIs and cannot quickly jump to the definition.

A semicolon is added to the end of above RTE_EXPORT_EXPERIMENTAL_SYMBOL
in this commit.

Signed-off-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>