net/zxdh: fix building check for arm64
The comment and docs say this should be built on x86_64 and aarch64,
but RTE_ARCH_64 is not aarch64, that's RTE_ARCH_ARM64, so fix it.

Also the check makes the build skip unless RTE_ARCH_X86_64 is true,
so switch it to an 'and' instead of an 'or', to build on both arm64
and x86_64, and skip elsewhere.

Fixes: 29e89288ff14 ("net/zxdh: add driver skeleton")
Cc: stable@dpdk.org

Signed-off-by: Luca Boccassi <luca.boccassi@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>