common/sfc_efx/base: use saved netport FEC choice by default
When the client driver starts the port, it invokes 'efx_port_init' API. At
this stage, libefx has no way of knowing whether the driver will prefer to
keep the existing link or revise, say, FEC mode. So if the current link is
up, it is better to leave it alone. This way, if the driver does not tweak
FEC mode with 'efx_phy_adv_cap_set' later on, the link will likely stay up.

Save the last FEC choice (select 'AUTO' on attach) and apply it by default.

Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Signed-off-by: 0-day Robot <robot@bytheb.org>