mbuf: record mbuf operations history
This feature is designed to monitor the lifecycle of mbufs
as they move between the applicationand the PMD.

It will allow us to track the operations and transitions
of each mbuf  throughout the system, helping in debugging
and understanding objects flow.

The implementation uses a dynamic field to store a 64-bit
history value in each mbuf. Each operation is represented
by a 4-bit value, allowing up to 16 operations to be tracked
per mbuf. The dynamic field is automatically initialized
when the first mbuf pool is created.

Signed-off-by: Shani Peretz <shperetz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>