net/idpf: add IDPF PCI class ID support
Current IDPF supports only the MEV device IDs. MMG has new set of device
IDs and same might be the case for the future devices. Instead of adding
new device IDs every time, make use of the IDPF PCI class ID(0x20001) to
differentiate between PF and VF.

Write and read the VF_ARQBAL register to find if the current device is
a PF or a VF.

Signed-off-by: Vemula Venkatesh <venkatesh.vemula@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>