drivers/net: update to remove queue stats from eth stats
Now that we have removed the queue stats from the ethdev stats, we need
to update the various PMDs to take account of the changes.

Update each stats_get function to take the extra parameters for qstats,
just marking it as unused if the driver does not fill in queue stats.
For those that do complete queue stats, update the structure reference
as appropriate.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>