eal: avoid using pthread_cancel
The multi-process handling thread can be managed more safely
by closing the underlying socket rather than pthread_cancel.
Closing the socket in the main thread will cause the read()
in the MP handler to unblock.

The use of atomic for updating the socket fd is not needed.
The fd for the socket is set before the thread is created
and not modified again until handler has exited.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>