app/testpmd: use table ID for jump to matcher action
Current implementation requires specifying the pointer to the table
you want to jump to in the jump to matcher action. It is inconvenient
since there is no table pointer shown anywhere in the table management.
Table creation/destruction uses the standard table ID for that purpose.
Use the table ID in the jump to matcher action as well.

Signed-off-by: Alexander Kozyrev <akozyrev@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>