build: factorize headers installation
Gather all headers in global variables, put headers installation
in a single location and make the headers check use them instead of
dpdk_chkinc_headers.

This rework reveals a number of issues:
- net/txgbe driver was directly installing its header, bypassing
  drivers/meson.build and skipping the headers check,
- arch headers were not checked except for x86,
- some driver headers were not checked,

For the last point, a build warning must be disabled (like it is
globally in DPDK) for headers from raw drivers.

Finally, checkpatches is updated to avoid reintroductions of
install_headers().

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>