net/cpfl: fix display of commands in the docs
The literal text of the commands to use were not displayed properly
because of a double ":" at the end of the preceding text, meaning that
the ".. code-block" marker got recognised as the literal text to
include. Change the double colon to single, and also indent the line
continuation in the multi-line flow-create command.

Fixes: 03f976012304 ("net/cpfl: adapt FXP to flow engine")
Cc: stable@dpdk.org

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>