net/enetfec: handle cache for forwarded packets
Previously, only cache flush (`dc cvac`) was used for packet buffers,
which could result in stale data being read when buffers were reused.
This patch replaces it with flush and invalidate (`dc civac`) to ensure
data consistency for forwarded packets.

Termination packets are excluded from this change as they are not freed
by the driver and do not require cache invalidation.

Signed-off-by: Hemant Agrawal <hemant.agrawal@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>