argparse: support parameters to short options without "="
When passing an optional parameter to a argparse argument, the only way
to do so has been to pass the argument using "=arg" on the end of the
option, for example "--foo=bar". However, for short options, i.e. those
of the form "-x", this requirement is unusual as getopt also supports
taking additional characters after the option as the parameter, e.g.
"-xarg".

Add support for this form of parameter for short options to argparse.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>