net/intel: remove redundant Rx offload check
Four Intel drivers use a common vector compatibility function to
determine if certain characteristics about a given rxq are compatible
with vector rx paths. The function checks if the buffer split offload is
enabled and disqualifies the rxq if it is, because that offload is not
available on any vector path. However, this check is redundant because
three of the drivers that use that function now use the common rx path
selection framework which performs essentially the same validation when
ensuring the requested offloads are available on the rx path to be
selected. The fourth driver does not support buffer split at all, so the
rx queue will never have that offload enabled as initialisation would
fail before getting to the path selection step. The redundant check is
removed in this commit.

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>